gitlab: merge files with more concise file name
